cmd/tailscale/tslockjsonv1: extract handling of TKA types
This patch extracts the handling of the TKA types for the
`tailscale lock` subcommand into its own tka.go file.

Updates #17613
