+++ cmd/tailscale/tslockjsonv1/tailnet-lock-status.go
@@ -6,9 +6,7 @@
 package tslockjsonv1
 
 import (
-	"encoding/base64"
 	jsonv1 "encoding/json"
-	"fmt"
 	"io"
 
 	"tailscale.com/cmd/tailscale/jsonoutput"
@@ -102,17 +100,6 @@ func toTailnetLockEnabledStatusV1(status *ipnstate.TailnetLockStatus) tailnetLoc
 	return out
 }
 
-// toTKAKeyV1 converts an [ipnstate.TKAKey] to the JSON output returned
-// by the CLI.
-func ipnTKAKeytoTKAKeyV1(key *ipnstate.TKAKey) tkaKeyV1 {
-	return tkaKeyV1{
-		Kind:   key.Kind,
-		Votes:  key.Votes,
-		Public: key.Key.CLIString(),
-		Meta:   key.Metadata,
-	}
-}
-
 type tailnetLockStatusV1Base struct {
 	// Enabled is true if Tailnet Lock is enabled.
 	Enabled bool
@@ -161,90 +148,4 @@ type tailnetLockEnabledStatusV1 struct {
 	// generated upon enablement. This field is empty if Tailnet Lock
 	// is disabled.
 	StateID uint64 `json:"State,omitzero"`
-}
-
-// tkaPeerV1 is the JSON representation of an [ipnstate.TKAPeer], which describes
-// a peer and its Tailnet Lock details.
-type tkaPeerV1 struct {
-	// Stable ID, i.e. [tailcfg.StableNodeID]
-	ID string
-
-	// DNS name
-	DNSName string
-
-	// Tailscale IP(s) assigned to this node
-	TailscaleIPs []string
-
-	// The node's public key
-	NodeKey string
-}
-
-// tkaPeerV1 is the JSON representation of a trusted [ipnstate.TKAPeer], which
-// has a node key signature.
-type tkaTrustedPeerV1 struct {
-	tkaPeerV1
-
-	// The node's key signature
-	NodeKeySignature *tkaNodeKeySignatureV1 `json:"NodeKeySignature,omitzero"`
-}
-
-func toTKAPeerV1(peer *ipnstate.TKAPeer) tkaPeerV1 {
-	out := tkaPeerV1{
-		DNSName: peer.Name,
-		ID:      string(peer.StableID),
-	}
-	for _, ip := range peer.TailscaleIPs {
-		out.TailscaleIPs = append(out.TailscaleIPs, ip.String())
-	}
-	out.NodeKey = peer.NodeKey.String()
-
-	return out
-}
-
-// tkaNodeKeySignatureV1 is the JSON representation of a [tka.NodeKeySignature],
-// which describes a signature that authorizes a specific node key.
-type tkaNodeKeySignatureV1 struct {
-	// SigKind identifies the variety of signature.
-	SigKind string
-
-	// PublicKey identifies the key.NodePublic which is being authorized.
-	// SigCredential signatures do not use this field.
-	PublicKey string `json:"PublicKey,omitzero"`
-
-	// KeyID identifies which key in the tailnet key authority should
-	// be used to verify this signature. Only set for SigDirect and
-	// SigCredential signature kinds.
-	KeyID string `json:"KeyID,omitzero"`
-
-	// Signature is the packed (R, S) ed25519 signature over all other
-	// fields of the structure.
-	Signature string
-
-	// Nested describes a NodeKeySignature which authorizes the node-key
-	// used as Pubkey. Only used for SigRotation signatures.
-	Nested *tkaNodeKeySignatureV1 `json:"Nested,omitzero"`
-
-	// WrappingPubkey specifies the ed25519 public key which must be used
-	// to sign a Signature which embeds this one.
-	WrappingPublicKey string `json:"WrappingPublicKey,omitzero"`
-}
-
-func toTKANodeKeySignatureV1(sig *tka.NodeKeySignature) *tkaNodeKeySignatureV1 {
-	out := tkaNodeKeySignatureV1{
-		SigKind: sig.SigKind.String(),
-	}
-	if len(sig.Pubkey) > 0 {
-		out.PublicKey = fmt.Sprintf("tlpub:%x", sig.Pubkey)
-	}
-	if len(sig.KeyID) > 0 {
-		out.KeyID = fmt.Sprintf("tlpub:%x", sig.KeyID)
-	}
-	out.Signature = base64.URLEncoding.EncodeToString(sig.Signature)
-	if sig.Nested != nil {
-		out.Nested = toTKANodeKeySignatureV1(sig.Nested)
-	}
-	if len(sig.WrappingPubkey) > 0 {
-		out.WrappingPublicKey = fmt.Sprintf("tlpub:%x", sig.WrappingPubkey)
-	}
-	return &out
 }
